+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * https://github.com/malachi-iot/estdlib/issues/139 fix `function_default` signature glitch
 * https://github.com/malachi-iot/estdlib/issues/140 AVR/UB fixups for ESTD_FLAGS
 * https://github.com/malachi-iot/estdlib/issues/142 scrubbing UTF8 -> ASCII
+* https://github.com/malachi-iot/estdlib/issues/143 removing legacy toString/fromString
 
 # v0.8.9 - 29JUL25